backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# AI Integration
openai>=1.6.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and logging
structlog==23.2.0

# Environment
python-dotenv==1.0.0

# Utilities
jinja2==3.1.2